feat: adds default config for avn pallet
